Update image_processing.py
Uses own draw_box function for placing the text information of detected objects into a box above the rectangle using a slightly bigger font. The fontfiles (10x20.pil and 10x20.pbm have to be present in the custom_components/deepstack_object/ directory. If the fontfiles are not found, the (small) default font is used.